+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to Notice Tracker will be documented in this file.
+All notable changes to Quietboard Notice Manager will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2026-03-05
 
 ### Added
-- Initial release of Notice Tracker
+- Initial release of Quietboard Notice Manager
 - Notice capture system with output buffering
 - Notice classification by type (success, error, warning, info, system, other)
 - Notice storage using WordPress options API